Fixed synch issues and sending of exercises back and forth  when creating a workout

+++ lib/database/Initialization/get_data_from_server.dart
@@ -4,10 +4,17 @@ import 'package:fitnessapp_idata2503/database/tables/exercise.dart';
 class GetDataFromServer {
   final ExerciseDao exerciseDao = ExerciseDao();
 
-  Future<void> syncExercises(String userId) async {
+  Future<void> syncExercises(String? userId) async {
+    if (userId == null) {
+      print("User ID is null");
+      return;
+    }
+    print("Syncing exercises for user: $userId");
     // Fetch all exercises from Firebase
     Map<String, dynamic> firebaseData = await exerciseDao.fetchAllExercisesFromFireBase(userId);
     List<Exercises> allExercises = firebaseData['exercises'];
+    print("All exercises: $allExercises");
+    print(allExercises[0].exerciseId);
 
     // Truncate the local database
     await exerciseDao.truncate();
